fix(abi): DbSetOrder(0) natural order local+remote; working 32-bit ace32.dll for rddads; e2e regression PRG
- DbSetOrder(0) parks the active order instead of keeping order 1:
  local branch of AdsSetIndexOrderByHandle(h,0) parks it, remote sends
  set_order_by_name("") so the server parks it too. Reported by
  Pritpal Bedi (Browse() after dbSetOrder(0) showed "name" order).
- AdsOpenIndex refresh re-registers opened index handles from the
  refreshed list (old_handles reuse); SetOrder after reindex no longer
  fails 5000 on a stale wire handle.
- AdsGetIndexOrderByHandle gains a remote branch: OrdNumber() returned
  0 for every remote table.
- ace32.dll exported only undecorated names, so every 32-bit rddads
  app died with 0xC0000139 at startup; it now exports all 360
  __stdcall-decorated (_AdsXxx@N) names via generated wrappers
  (src/abi/ace_stdcall_x86.c) and openads_ace_x86_stdcall.def is the
  default .def for 32-bit builds. AdsSetIndexOrderByHandle added to
  the x86 export set.
- ADSHANDLE was uint64_t while SAP ACE defines it 32-bit: every
  ADSHANDLE* out-param wrote 8 bytes into 32-bit callers' 4-byte
  storage and crashed them inside the first AdsConnect60. Now
  uint32_t, with explicit narrowing at the ABI boundary.
- dist import libs for x86 regenerated for the @N export surface
  (MSVC: DLL link byproduct; MinGW: dlltool --no-leading-underscore;
  Borland: implib from the DLL). gen.ps1 picks the freshest build tree.
- tests/e2e/b_big_e2e.prg: comprehensive end-to-end regression PRG
  (21 sections: version, FIELD-> strip, bag auto-open, dbSetIndex,
  dbSetOrder(0)/n, scope+deleted walk/KeyNo, custom-ext bags, reindex,
  9-record commit timing, second connection, 3 threads). build_e2e.bat
  builds both bitnesses; 21/21 PASS on x64+x86, local+remote.
  Requested by Pritpal Bedi as the pre-release gate.

v1.8.46

+++ include/openads/ace.h
@@ -34,7 +34,11 @@ typedef uint32_t UNSIGNED32;
 typedef int32_t  SIGNED32;
 typedef int64_t  SIGNED64;
 typedef uint64_t UNSIGNED64;
-typedef uint64_t ADSHANDLE;
+// SAP ACE defines ADSHANDLE as a 32-bit value on every platform.
+// Declaring it 64-bit here made every ADSHANDLE-by-reference out
+// parameter write 8 bytes into the caller's 4-byte storage, which
+// corrupts the stack of 32-bit callers (rddads) — v1.8.46.
+typedef uint32_t ADSHANDLE;
 
 #define ADS_DEFAULT 0
 #define ADS_NTX     1
